fix: Bug fixes and improvements
