+++ src/plugin/plugin.ts
@@ -24,8 +24,8 @@ import { getPty } from '../pty/index.js';
 import { SequentialPty } from '../pty/seqeuntial-pty.js';
 import { Resource } from '../resource/resource.js';
 import { ResourceController } from '../resource/resource-controller.js';
-import { VerbosityLevel } from '../utils/internal-utils.js';
 import { ptyLocalStorage } from '../utils/pty-local-storage.js';
+import { VerbosityLevel } from '../utils/verbosity-level.js';
 
 export class Plugin {
   planStorage: Map<string, Plan<any>>;
