+++ src/utils/verbosity-level.ts
@@ -1,0 +1,11 @@
+export const VerbosityLevel = new class {
+  level = 0;
+
+  get() {
+    return this.level;
+  }
+
+  set(level: number) {
+    this.level = level;
+  }
+}
